requirements.txt
# Croniter package for djangoq
croniter==1.3.8

# Request Authentication Caching
cryptography==38.0.3
future==0.18.3

# Django and Django REST Framework
Django==4.1.2
django-cors-headers==4.3.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
django-db-geventpool==4.0.1
django-request-logging==0.7.5
django-filter==21.1

# DjangoQ for running async tasks
django-q==1.3.9

# Read Database Credentials as URL
dj-database-url==0.5.0

# API server
gevent==23.9.1
gunicorn==20.1.0

# Platform SDK
fyle==0.36.0

# Business central sdk
ms-dynamics-business-central-sdk==1.3.2

# Reusable Fyle Packages
fyle-rest-auth==1.5.0
fyle-accounting-mappings==1.27.3
fyle-integrations-platform-connector==1.36.1

# Postgres Dependincies
psycopg2-binary==2.9.9

# Pylint
pylint==2.7.4

# Date Parser
python-dateutil==2.8.1

# Pylint Related Libraries
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-mock==3.8.2

# Sendgrid for sending emails_selected
sendgrid==6.9.7
sentry-sdk==1.19.1
